Updating project structure, moving away from a monolith. Updating tests accordingly
