+++ tests/test_auxillary_functions.py
@@ -2,6 +2,7 @@ import os
 import unittest
 
 import todo_or_not.todo_check
+import todo_or_not.utility
 
 
 class TestLocalization(unittest.TestCase):
@@ -11,13 +12,13 @@ class TestLocalization(unittest.TestCase):
 
     def test_unsupported_and_supported_regions(self):
         # Check unsupported
-        region = todo_or_not.todo_check.get_region()
+        region = todo_or_not.utility.get_region()
 
         self.assertEqual(region, "en_us")
 
         # Check supported
         os.environ["REGION"] = "ko_kr"
-        region = todo_or_not.todo_check.get_region()
+        region = todo_or_not.utility.get_region()
 
         self.assertEqual(region, "ko_kr")
 
@@ -25,13 +26,13 @@ class TestLocalization(unittest.TestCase):
 
     def test_unsupported_and_supported_operating_systems(self):
         # Check unsupported
-        region = todo_or_not.todo_check.get_os()
+        region = todo_or_not.utility.get_os()
 
         self.assertEqual(region, "default")
 
         # Check supported
         os.environ["OS"] = "windows_nt"
-        region = todo_or_not.todo_check.get_os()
+        region = todo_or_not.utility.get_os()
 
         self.assertEqual(region, "windows_nt")
