+++ todo_or_not/todo_check.py
@@ -1,5 +1,4 @@
 import glob
-import hashlib
 import json
 import os
 import subprocess
@@ -11,352 +10,29 @@ from tqdm import tqdm
 from typer import run
 from typing_extensions import Annotated
 
-import todo_or_not  # todoon
+import todo_or_not.utility as util
 from todo_or_not.localize import LOCALIZE  # todoon
 from todo_or_not.localize import SUPPORTED_ENCODINGS_TODOIGNORE  # todoon
 from todo_or_not.localize import SUPPORTED_ENCODINGS_TODO_CHECK  # todoon
+# from todo_or_not.todo_grammar import find_language, TodoGrammar
+from todo_or_not.todo_hit import Hit
 
 todoon_app = typer.Typer(name="todoon")  # todoon
 
-LOG_LEVEL_NONE = 0
-LOG_LEVEL_SUMMARY_ONLY = 1
-LOG_LEVEL_NORMAL = 2
-LOG_LEVEL_VERBOSE = 3
 
-
-def print_wrap(
-    msg: str, msg_level=LOG_LEVEL_NORMAL, log_level=LOG_LEVEL_NORMAL, file=sys.stdout
-):
-    if msg_level <= log_level:
-        print(msg, file=file)
-
-
-def version_callback(log_level=LOG_LEVEL_NORMAL):
-    print_wrap(
-        log_level=log_level,
-        msg=f"TODO-Or-Not v{todo_or_not.__version__} ({todo_or_not.version_date})",  # todoon
-    )
-    exit(0)
-
-
-def get_project_dir():
-    return os.getcwd()
-
-
-def get_todo_ignore_path():  # todoon
-    return os.path.join(get_project_dir(), ".todo-ignore")  # todoon
-
-
-def get_is_debug():
-    _debug = os.environ.get("DEBUG", "False")
-    if _debug == "True":
-        return True
-    else:
-        return False
-
-
-def get_max_issues():
-    _max_issues = os.environ.get("MAXIMUM_ISSUES_GENERATED", "8")
-
-    try:
-        max_issues = int(_max_issues)
-    except ValueError:
-        max_issues = 8
-
-    return max_issues
-
-
-def get_pertinent_line_limit():
-    _pertinent_line_limit = os.environ.get("PERTINENT_LINE_LIMIT", "8")
-
-    try:
-        pertinent_line_limit = int(_pertinent_line_limit)
-    except ValueError:
-        pertinent_line_limit = 8
-
-    return pertinent_line_limit
-
-
-def get_region(log_level=LOG_LEVEL_NORMAL):
-    region = os.environ.get("REGION", "en_us")
-
-    # Validate that we support the region, otherwise default to something we have
-    if region not in LOCALIZE:
-        print_wrap(
-            log_level=log_level,
-            msg=f"{LOCALIZE['en_us']['warning_using_default_region']} region",
-            file=sys.stderr,
-        )
-        region = "en_us"
-
-    return region
-
-
-def get_os(log_level=LOG_LEVEL_NORMAL):
-    _os = os.environ.get("OS", "default")
-    _os = _os.lower()
-
-    # Validate that we support the region, otherwise default to something we have
-    if _os not in LOCALIZE:
-        print_wrap(
-            log_level=log_level,
-            msg=f"{LOCALIZE[get_region()]['warning_using_default_os']} _os",
-            file=sys.stderr,
-        )
-        _os = "default"
-
-    return _os
-
-
-class Hit:
-    def __init__(
-        self,
-        source_file: str,
-        source_line: int,
-        found_keys: list[str],
-        pertinent_lines: list[str],
-        trigger_line_index: int,
-    ):
-        self.found_keys = found_keys
-        self.source_file = source_file
-        self.source_line = source_line
-        self.pertinent_lines = pertinent_lines
-        self.trigger_line_index = trigger_line_index
-
-        self.structured_title = None
-        self.structured_body = None
-        self.structured_labels = None
-
-        # If this is a structured comment, parse out the title, body, and labels
-        if "|" in self.pertinent_lines[trigger_line_index]:
-            head, body = self.pertinent_lines[trigger_line_index].split("|", 1)
-
-            self.structured_title = head.strip()
-            self.structured_body = body.strip()
-
-            # If there is an # in the body, there may be labels to find
-            if "#" in self.structured_body:
-                self.structured_labels = []
-
-                _potential_tags = self.structured_body.split("#")
-
-                # Skip the first item in this list, because the labels will come after the #
-                for _potential in _potential_tags[1:]:
-                    _label = _potential.split(" ", 1)[0]
-
-                    if len(_label) > 0:
-                        self.structured_labels.append(_label)
-
-                if len(self.structured_labels) == 0:
-                    self.structured_labels = None
-
-    def __repr__(self):
-
-        if self.structured_title is not None:
-            return self.structured_title
-
-        _line = self.get_triggering_line()
-        _line_number = self.get_line_number()
-        _found_keys = self.get_found_keys()
-
-        header = f"[{self.get_found_keys()}]"
-        _pad = 16 - len(header)
-        padding = " " * _pad
-        header = header + padding
-
-        location = os.path.relpath(self.source_file, get_project_dir())
-        location = f"{location}:{_line_number}"
-        _pad = 16 - len(location)
-        padding = " " * _pad
-        location = location + padding
-
-        return f"{header} - {location} - {_line.strip()}"
-
-    def __eq__(self, other):
-        if not isinstance(other, Hit):
-            # don't attempt to compare against unrelated types
-            return NotImplemented
-
-        return (
-            self.found_keys == other.found_keys
-            and self.source_file == other.source_file
-            and self.source_line == other.source_line
-            and self.pertinent_lines == other.pertinent_lines
-            and self.trigger_line_index == other.trigger_line_index
-            and self.structured_title == other.structured_title
-            and self.structured_body == other.structured_body
-            and self.structured_labels == other.structured_labels
-        )
-
-    def get_triggering_line(self):
-        return self.pertinent_lines[self.trigger_line_index]
-
-    def get_line_number(self):
-        return self.source_line
-
-    def get_found_keys(self):
-        return ", ".join(self.found_keys).upper()
-
-    def get_file_extension(self):
-        return self.source_file.split(".")[-1:][0]
-
-    def get_pertinent_lines(self):
-        starting_line_number = self.source_line - self.trigger_line_index
-        _max_line = self.source_line + (
-            len(self.pertinent_lines) - self.trigger_line_index
-        )
-
-        def _parse_line_number(_l: int, star: bool = False) -> str:
-            _padding = len(str(_max_line))
-            _padding += 3
-
-            _decoration = "* " if star else "  "
-
-            return f"{_decoration}{str(_l)}:".rjust(_padding, " ")
-
-        output = f"```{self.get_file_extension()}\n"
-
-        for pertinent_line in self.pertinent_lines:
-            output += (
-                f"{_parse_line_number(starting_line_number, starting_line_number == self.source_line)}\t"
-                f"{pertinent_line}"
-            )
-            starting_line_number += 1
-
-        output += "```"
-
-        return output
-
-    def get_title(self):
-        return (
-            self.generic_title()
-            if self.structured_title is None
-            else self.structured_title
-        )
-
-    def generic_title(self):
-        return f"{self.get_found_keys()} - {self.get_triggering_line()}"
-
-    def generate_issue(
-        self, _test: bool = False, log_level=LOG_LEVEL_NORMAL
-    ) -> str or bool:
-
-        repo_uri = f"https://github.com/None"
-
-        github_ref = "$NONE"
-        triggered_by = "$NONE"
-        owner, repo = "$NONE", "$NONE"
-
-        if not (get_is_debug() or _test):
-            repo_uri = f"https://github.com/{os.environ.get('GITHUB_REPOSITORY')}"
-
-            github_ref = os.environ.get("GITHUB_REF_NAME", "$NONE")
-            triggered_by = os.environ.get("GITHUB_TRIGGERING_ACTOR", "$NONE")
-            owner, repo = os.environ.get("GITHUB_REPOSITORY", "$NONE/$NONE").split("/")
-
-            # Find missing env variables
-            missing_envs = []
-
-            if github_ref == "$NONE":
-                print_wrap(
-                    log_level=log_level,
-                    msg=f"{LOCALIZE[get_region()]['error_no_env']}: GITHUB_REF_NAME",
-                    file=sys.stderr,
-                )
-                missing_envs.append("GITHUB_REF_NAME")
-            if triggered_by == "$NONE":
-                print_wrap(
-                    log_level=log_level,
-                    msg=f"{LOCALIZE[get_region()]['error_no_env']}: GITHUB_TRIGGERING_ACTOR",
-                    file=sys.stderr,
-                )
-                missing_envs.append("GITHUB_TRIGGERING_ACTOR")
-            if owner == "$NONE":
-                print_wrap(
-                    log_level=log_level,
-                    msg=f"{LOCALIZE[get_region()]['error_no_env']}: GITHUB_REPOSITORY",
-                    file=sys.stderr,
-                )
-                missing_envs.append("GITHUB_REPOSITORY")
-            if repo == "$NONE":
-                print_wrap(
-                    log_level=log_level,
-                    msg=f"{LOCALIZE[get_region()]['error_no_env']}: GITHUB_REPOSITORY",
-                    file=sys.stderr,
-                )
-                missing_envs.append("GITHUB_REPOSITORY")
-
-            if len(missing_envs) > 0:
-                return False
-
-        reference_file = self.source_file.split(":")[0]
-
-        reference_uri = f"{repo_uri}/blob/{github_ref}/{reference_file}"
-
-        body = (
-            f"## {self if self.structured_body is None else self.structured_body}\n\n"
-            f"{self.get_pertinent_lines()}\n\n"
-            f"{LOCALIZE[get_region()]['issue_body_reference_link']}: <a href=\"{reference_uri}\">{self.source_file}</a>"
-        )
-
-        # Sanitize @ to prevent abuse
-        body.replace("@", "@<!-- -->")
-
-        api_call = [
-            "gh",
-            "api",
-            "--method",
-            "POST",
-            "-H",
-            "Accept: application/vnd.github+json",
-            "-H",
-            "X-GitHub-Api-Version: 2022-11-28",
-            f"/repos/{owner}/{repo}/issues",
-            "-f",
-            f"title={self.get_title()}",
-            "-f",
-            f"body={body}",
-            "-f",
-            f"assignees[]={triggered_by}",
-        ]
-
-        if self.structured_labels is not None:
-            for label in self.structured_labels:
-                api_call.append("-f")
-                api_call.append(f"labels[]={label}")
-
-        if not (get_is_debug() or _test):
-            try:
-                _output = subprocess.check_output(api_call)
-            except subprocess.CalledProcessError as e:
-                print_wrap(log_level=log_level, msg=str(e), file=sys.stderr)
-                _output = False
-        else:
-            _output = True
-            print_wrap(log_level=log_level, msg=str(api_call))
-
-        return _output
-
-
-def _hash(hit_str: str):
-    m = hashlib.sha1()
-    m.update(bytes(hit_str, "utf-8"))
-    return m.hexdigest()
-
-
-def find_lines(
+def find_hits(
     filename: str,
     verbose: bool,
     ignore_flag: str,
-    ignore_keys: list,
-    log_level=LOG_LEVEL_NORMAL,
+    parsers: dict,
+    log_level=util.LOG_LEVEL_NORMAL,
 ) -> tuple[list[Hit], str or None]:
     """
     Finds and returns each line of a file that contains a key
     :param verbose: Print lengthy feedback which includes (encoding failures)
     :param ignore_flag: The flag which, when detected on a triggering line, will ignore that line
     :param filename: File to open() read-only
-    :param ignore_keys: Keys to check each line for
+    :param parsers: Parsers that have been built for discovered languages
     :param log_level: The importance of any feedback prints (e.g. 0=NONE, 3=VERBOSE)
     :return:
      | List of lines of text and their line number that contain at least one key and the keys each contains
@@ -368,6 +44,17 @@ def find_lines(
 
     if use_encoding is not None:
         with open(filename, "r", encoding=use_encoding) as file:
+            # Get the extension of this file and parse its language, note that several extensions may be associated
+            # with a single language, so we must find the language common to those extensions.
+        # file_extension = filename.rsplit(".", 1)[-1]
+        # file_language = find_language(file_extension)
+
+            # If that language does not yet have a parser built, we must build one (note that calling the TodoGrammar
+            # constructor with the file_extension will functional identically between different file extensions for
+            # the same language)
+        # if file_language not in parsers.keys():
+        #     parsers[file_language] = TodoGrammar(file_extension)
+
             line_number = 0
             lines = file.readlines()
 
@@ -381,18 +68,22 @@ def find_lines(
                 # Collect the found keys and their associated info
                 _found_keys = []
 
-                for key in ignore_keys:
+                for key in ["todo", "fixme"]:
                     if key.lower() in _line.lower():
                         _found_keys.append(key)
 
                 if len(_found_keys) > 0:
+            # _use_parser = parsers[file_language]
+            # _potential_hit = _use_parser.parse(_line)
+
+            # if False:
                     # Collect surrounding lines that may be pertinent
                     _pertinent_lines = []
 
                     # Look at lines before the pertinent line
                     _i = line_number - 1
                     while (
-                        abs(line_number - _i) <= get_pertinent_line_limit() and _i >= 0
+                        abs(line_number - _i) <= util.get_pertinent_line_limit() and _i >= 0
                     ):
                         _i -= 1
                         if len(lines[_i].strip()) > 0:
@@ -407,7 +98,7 @@ def find_lines(
 
                     # Look at lines after the pertinent line
                     _i = line_number
-                    while abs(_i - line_number) <= get_pertinent_line_limit():
+                    while abs(_i - line_number) <= util.get_pertinent_line_limit():
                         if _i < len(lines) and len(lines[_i].strip()) > 0:
                             _pertinent_lines.append(lines[_i])
                         else:
@@ -416,8 +107,9 @@ def find_lines(
                         _i += 1
 
                     _hit = Hit(
-                        os.path.relpath(filename, get_project_dir()),
+                        os.path.relpath(filename, os.getcwd()),
                         line_number,
+                        # {"status": "I'm not done yet"},
                         _found_keys,
                         _pertinent_lines,
                         _trigger_line,
@@ -425,10 +117,10 @@ def find_lines(
                     output.append(_hit)
 
     else:
-        print_wrap(
+        util.print_wrap(
             log_level=log_level,
-            msg_level=LOG_LEVEL_VERBOSE,
-            msg=f"{LOCALIZE[get_region()]['warning_encoding_not_supported']} \n * {filename}",
+            msg_level=util.LOG_LEVEL_VERBOSE,
+            msg=f"{LOCALIZE[util.get_region()]['warning_encoding_not_supported']} \n * {filename}",
         )
 
     return output, use_encoding
@@ -454,7 +146,7 @@ def paste_contents_into_file(other_file_names: list[str], target_file: TextIO):
 
 
 def get_bot_submitted_issues(
-    _test: bool = False, log_level=LOG_LEVEL_NORMAL
+    _test: bool = False, log_level=util.LOG_LEVEL_NORMAL
 ) -> list[dict] or bool:
     """
     Makes a gh cli request for all issues submitted by app/todo-or-not, parses them, and returns them as a # todoon
@@ -464,12 +156,12 @@ def get_bot_submitted_issues(
     owner, repo = "owner", "repository"
 
     try:
-        if not (get_is_debug() or _test):
+        if not (util.get_is_debug() or _test):
             owner, repo = os.environ.get("GITHUB_REPOSITORY").split("/")
     except AttributeError as _:
-        print_wrap(
+        util.print_wrap(
             log_level=log_level,
-            msg=f"{LOCALIZE[get_region()]['error_no_env']}: GITHUB_REPOSITORY",
+            msg=f"{LOCALIZE[util.get_region()]['error_no_env']}: GITHUB_REPOSITORY",
             file=sys.stderr,
         )
 
@@ -483,11 +175,11 @@ def get_bot_submitted_issues(
         f"/repos/{owner}/{repo}/issues?creator=app%2Ftodo-or-not&state=all",  # todoon
     ]
 
-    if not (get_is_debug() or _test):
+    if not (util.get_is_debug() or _test):
         try:
             response = subprocess.check_output(query)
         except subprocess.CalledProcessError as e:
-            print_wrap(log_level=log_level, msg=str(e), file=sys.stderr)
+            util.print_wrap(log_level=log_level, msg=str(e), file=sys.stderr)
             return False
 
         _str = response.decode("utf-8")
@@ -495,12 +187,12 @@ def get_bot_submitted_issues(
 
         return json.loads(_str)
     else:
-        print_wrap(log_level=log_level, msg=str(query), file=sys.stderr)
+        util.print_wrap(log_level=log_level, msg=str(query), file=sys.stderr)
         return False
 
 
 def get_encoding(
-    _target_path: str, _supported_encodings: list[str], log_level=LOG_LEVEL_NORMAL
+    _target_path: str, _supported_encodings: list[str], log_level=util.LOG_LEVEL_NORMAL
 ) -> str or None:
     """
     :param _target_path: A path-like string pointing to the file for which we want to get a valid encoding
@@ -511,9 +203,9 @@ def get_encoding(
     try:
         assert os.path.isfile(_target_path)
     except AssertionError:
-        print_wrap(
+        util.print_wrap(
             log_level=log_level,
-            msg=f"{LOCALIZE[get_region()]['error_is_not_file']}: {_target_path}",
+            msg=f"{LOCALIZE[util.get_region()]['error_is_not_file']}: {_target_path}",
             file=sys.stderr,
         )
         return None
@@ -596,13 +288,13 @@ def todoon(  # todoon
     ignored_files = []
     ignored_dirs = []
 
-    log_level = LOG_LEVEL_NORMAL
+    log_level = util.LOG_LEVEL_NORMAL
     if verbose:
-        log_level = LOG_LEVEL_VERBOSE
+        log_level = util.LOG_LEVEL_VERBOSE
     if print_summary_only:
-        log_level = LOG_LEVEL_SUMMARY_ONLY
+        log_level = util.LOG_LEVEL_SUMMARY_ONLY
     if print_nothing:
-        log_level = LOG_LEVEL_NONE
+        log_level = util.LOG_LEVEL_NONE
 
     use_specified_files = False
 
@@ -610,7 +302,7 @@ def todoon(  # todoon
         use_specified_files = True
 
     if version:
-        version_callback()
+        util.version_callback()
 
     #############################################
     # Handle settings
@@ -637,19 +329,19 @@ def todoon(  # todoon
             # Unless --force is specified,
             # a .todo-ignore in a supported encoding must be located at the project's top level # todoon
             use_encoding = get_encoding(
-                get_todo_ignore_path(), SUPPORTED_ENCODINGS_TODOIGNORE  # todoon
+                util.get_todo_ignore_path(), SUPPORTED_ENCODINGS_TODOIGNORE  # todoon
             )
 
             # If we weren't able to find a file in a supported encoding, program must exit
             if use_encoding is None:
-                print_wrap(log_level=log_level,
-                           msg=LOCALIZE[get_region()]["error_todo_ignore_not_supported"], file=sys.stderr  # todoon
+                util.print_wrap(log_level=log_level,
+                           msg=LOCALIZE[util.get_region()]["error_todo_ignore_not_supported"], file=sys.stderr  # todoon
                            )
                 exit(1)
 
             # ... actually do the reading of the .todo-ignore # todoon
             with open(
-                    get_todo_ignore_path(), "r", encoding=use_encoding  # todoon
+                    util.get_todo_ignore_path(), "r", encoding=use_encoding  # todoon
             ) as _ignore:
                 for line in _ignore.readlines():
                     if not line.startswith("#") and len(line) > 1:
@@ -658,7 +350,7 @@ def todoon(  # todoon
                         else:
                             cur_name = line
 
-                        cur_path = os.path.join(get_project_dir(), cur_name)
+                        cur_path = os.path.join(os.getcwd(), cur_name)
 
                         # Resolve wildcards
                         if '*' in cur_path:
@@ -671,8 +363,8 @@ def todoon(  # todoon
                             ignored_dirs.append(cur_path)
 
                 if len(ignored_files) == 0 and len(ignored_dirs) == 0:
-                    print_wrap(log_level=log_level,
-                               msg=LOCALIZE[get_region()][
+                    util.print_wrap(log_level=log_level,
+                               msg=LOCALIZE[util.get_region()][
                                    "warning_run_with_empty_todo_ignore"  # todoon
                                ],
                                file=sys.stderr,
@@ -681,9 +373,9 @@ def todoon(  # todoon
                 # Ignore the .todo-ignore itself # todoon
                 ignored_files.append(os.path.abspath(_ignore.name))
         else:
-            print_wrap(log_level=log_level,
-                       msg=f"{LOCALIZE[get_region()]['error_todo_ignore_not_found']}"  # todoon
-                           f"[{LOCALIZE[get_os()]['shell_sigint']}]",
+            util.print_wrap(log_level=log_level,
+                       msg=f"{LOCALIZE[util.get_region()]['error_todo_ignore_not_found']}"  # todoon
+                           f"[{LOCALIZE[util.get_os()]['shell_sigint']}]",
                        file=sys.stderr,
                        )
 
@@ -695,10 +387,10 @@ def todoon(  # todoon
     if not use_specified_files:
         os.environ["TODOON_STATUS"] = "collecting-targets"  # todoon
         # Ignore this script if in DEBUG
-        if get_is_debug():
+        if util.get_is_debug():
             ignored_files.append(__file__)
 
-        _walk = os.walk(get_project_dir(), topdown=True)
+        _walk = os.walk(os.getcwd(), topdown=True)
 
         for dirpath, dirnames, filenames in _walk:
             _to_remove = []
@@ -729,7 +421,7 @@ def todoon(  # todoon
     else:
         # Collect specified files
         for file in files:
-            current_path = os.path.join(get_project_dir(), file)
+            current_path = os.path.join(os.getcwd(), file)
 
             # If the specified path is a file, simply add it
             if os.path.isfile(current_path):
@@ -752,8 +444,8 @@ def todoon(  # todoon
     # Warning if issues options are used when issue mode is not enabled
     if print_mode:
         if fail_closed_duplicates:
-            print_wrap(log_level=log_level,
-                       msg=f"{LOCALIZE[get_region()]['warning_nonissue_mode_closed_duplicate_used']}", file=sys.stderr
+            util.print_wrap(log_level=log_level,
+                       msg=f"{LOCALIZE[util.get_region()]['warning_nonissue_mode_closed_duplicate_used']}", file=sys.stderr
                        )
 
     # When pinging for all queries their titles are hashed and saved here along with their state,
@@ -769,10 +461,10 @@ def todoon(  # todoon
 
         if todoon_created_issues is not False:  # todoon
             for issue in todoon_created_issues:  # todoon
-                existing_issues_hashed[_hash(issue["title"])] = issue["state"]
+                existing_issues_hashed[util._hash(issue["title"])] = issue["state"]
         else:
-            print_wrap(log_level=log_level,
-                       msg=f"{LOCALIZE[get_region()]['error_gh_issues_read_failed']}", file=sys.stderr
+            util.print_wrap(log_level=log_level,
+                       msg=f"{LOCALIZE[util.get_region()]['error_gh_issues_read_failed']}", file=sys.stderr
                        )
 
     #############################################
@@ -802,8 +494,8 @@ def todoon(  # todoon
     # TODO NEW Localization 'progress_bar_run_unit' | "file" #localization
     # TODO NEW Localization 'progress_bar_run_desc' | "scanning files" #localization
     if show_progress_bar:
-        _target_iterator = tqdm(targets, unit=LOCALIZE[get_region()]['progress_bar_run_unit'],
-                                desc=LOCALIZE[get_region()]['progress_bar_run_desc'])
+        _target_iterator = tqdm(targets, unit=LOCALIZE[util.get_region()]['progress_bar_run_unit'],
+                                desc=LOCALIZE[util.get_region()]['progress_bar_run_desc'])
 
     for target in _target_iterator:
 
@@ -811,8 +503,10 @@ def todoon(  # todoon
         _i += 1
         os.environ["TODOON_PROGRESS"] = str(round(_i / (len(targets)), 1))  # todoon
 
+        parsers = {}
+
         # Generate the hits for each target collected
-        hits, _enc = find_lines(target, verbose, "# todoon", ["todo", "fixme"], log_level=log_level)
+        hits, _enc = find_hits(target, verbose, "# todoon", parsers, log_level=log_level)
 
         if _enc is None:
             number_of_encoding_failures += 1
@@ -830,26 +524,26 @@ def todoon(  # todoon
                 # Special handling for the ISSUE mode
 
                 if not print_mode:
-                    _this_hit_hashed = _hash(hit.get_title())
+                    _this_hit_hashed = util._hash(hit.get_title())
 
                     # Check if the app already created this hit's title in open AND closed issues
                     if _this_hit_hashed not in existing_issues_hashed:
 
                         # Limit the number of issues created in one run
-                        if number_of_issues < get_max_issues():
+                        if number_of_issues < util.get_max_issues():
                             output = hit.generate_issue()
 
                             if output is not False:
                                 number_of_issues += 1
                             else:
-                                print_wrap(log_level=log_level,
-                                           msg=f"{LOCALIZE[get_region()]['error_gh_issues_create_failed']}",
+                                util.print_wrap(log_level=log_level,
+                                           msg=f"{LOCALIZE[util.get_region()]['error_gh_issues_create_failed']}",
                                            file=sys.stderr
                                            )
 
                         else:
-                            print_wrap(log_level=log_level,
-                                       msg=LOCALIZE[get_region()][
+                            util.print_wrap(log_level=log_level,
+                                       msg=LOCALIZE[util.get_region()][
                                            "error_exceeded_maximum_issues"
                                        ],
                                        file=sys.stderr,
@@ -857,15 +551,15 @@ def todoon(  # todoon
                             exit(1)
                     # If this title exists AND is closed, potentially fail the check
                     elif existing_issues_hashed[_this_hit_hashed] == "closed":
-                        print_wrap(log_level=log_level,
-                                   msg=f"{LOCALIZE[get_region()]['warning_duplicate_closed_issue']}: {hit}",
+                        util.print_wrap(log_level=log_level,
+                                   msg=f"{LOCALIZE[util.get_region()]['warning_duplicate_closed_issue']}: {hit}",
                                    file=sys.stderr
                                    )
                         number_of_closed_issues += 1
                     # If this title already exists, notify but do not halt
                     else:
-                        print_wrap(log_level=log_level,
-                                   msg=f"{LOCALIZE[get_region()]['info_duplicate_issue_avoided']}: {hit}",
+                        util.print_wrap(log_level=log_level,
+                                   msg=f"{LOCALIZE[util.get_region()]['info_duplicate_issue_avoided']}: {hit}",
                                    file=sys.stderr,
                                    )
                         number_of_duplicate_issues_avoided += 1
@@ -874,14 +568,14 @@ def todoon(  # todoon
                 # If not in ISSUE mode, print hit to stderr
 
                 else:
-                    print_wrap(log_level=log_level,
+                    util.print_wrap(log_level=log_level,
                                msg=str(hit), file=sys.stderr)
 
     #############################################
     # Summarize the run of todo-check  # todoon
     #############################################
 
-    summary = f"\n##########################\n# {LOCALIZE[get_region()]['summary_title']}\n"
+    summary = f"\n##########################\n# {LOCALIZE[util.get_region()]['summary_title']}\n"
     # Mode the tool was run in
     if print_mode:
         summary += "# (PRINT MODE)\n"
@@ -893,46 +587,46 @@ def todoon(  # todoon
 
     # Number of encoding failures
     if number_of_encoding_failures > 1:
-        summary += f"# {number_of_encoding_failures} {LOCALIZE[get_region()]['summary_encoding_unsupported_plural']}\n"
+        summary += f"# {number_of_encoding_failures} {LOCALIZE[util.get_region()]['summary_encoding_unsupported_plural']}\n"
     elif number_of_encoding_failures == 1:
-        summary += f"# {number_of_encoding_failures} {LOCALIZE[get_region()]['summary_encoding_unsupported_singular']}\n"
+        summary += f"# {number_of_encoding_failures} {LOCALIZE[util.get_region()]['summary_encoding_unsupported_singular']}\n"
 
     # Total number of files scanned
     if number_of_files_scanned > 1:
         summary += (f"# {number_of_files_scanned} "
-                    f"{LOCALIZE[get_region()]['summary_files_scanned_plural']}\n")
+                    f"{LOCALIZE[util.get_region()]['summary_files_scanned_plural']}\n")
     elif number_of_files_scanned == 1:
         summary += (f"# {number_of_files_scanned} "
-                    f"{LOCALIZE[get_region()]['summary_files_scanned_singular']}\n")
+                    f"{LOCALIZE[util.get_region()]['summary_files_scanned_singular']}\n")
 
         # Number of issues (if any) that were generated
     if not print_mode:
         # Total number of issues generated
         if number_of_issues > 1:
             summary += (f"# {number_of_issues} "
-                        f"{LOCALIZE[get_region()]['summary_issues_generated_plural']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_issues_generated_plural']}\n")
         elif number_of_issues == 1:
             summary += (f"# {number_of_issues} "
-                        f"{LOCALIZE[get_region()]['summary_issues_generated_singular']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_issues_generated_singular']}\n")
         else:
             summary += (f"# "
-                        f"{LOCALIZE[get_region()]['summary_issues_generated_none']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_issues_generated_none']}\n")
 
         # Total number of duplicate issues avoided
         if number_of_duplicate_issues_avoided > 1:
             summary += (f"# {number_of_duplicate_issues_avoided} "
-                        f"{LOCALIZE[get_region()]['summary_duplicate_issues_avoided_plural']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_duplicate_issues_avoided_plural']}\n")
         elif number_of_duplicate_issues_avoided == 1:
             summary += (f"# {number_of_duplicate_issues_avoided} "
-                        f"{LOCALIZE[get_region()]['summary_duplicate_issues_avoided_singular']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_duplicate_issues_avoided_singular']}\n")
 
         # Total number of duplicate closed issues
         if number_of_closed_issues > 1:
             summary += (f"# {number_of_closed_issues} "
-                        f"{LOCALIZE[get_region()]['summary_duplicate_closed_issues_plural']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_duplicate_closed_issues_plural']}\n")
         elif number_of_closed_issues == 1:
             summary += (f"# {number_of_closed_issues} "
-                        f"{LOCALIZE[get_region()]['summary_duplicate_closed_issues_singular']}\n")
+                        f"{LOCALIZE[util.get_region()]['summary_duplicate_closed_issues_singular']}\n")
 
     summary += "##########################\n\n"
 
@@ -940,17 +634,17 @@ def todoon(  # todoon
     if number_of_hits > 0:
         if silent:
             # TODO NEW Localization 'summary_found_issues_silent' | "INFO: New issues detected, but todoon ran in --silent mode" #localization
-            summary += f"  * {LOCALIZE[get_region()]['summary_found_issues_silent']}\n"
+            summary += f"  * {LOCALIZE[util.get_region()]['summary_found_issues_silent']}\n"
         else:
-            summary += f"  * {LOCALIZE[get_region()]['summary_fail_issues_no_silent']}\n"
+            summary += f"  * {LOCALIZE[util.get_region()]['summary_fail_issues_no_silent']}\n"
 
     if number_of_closed_issues > 0 and fail_closed_duplicates:
-        summary += f"  * {LOCALIZE[get_region()]['summary_fail_duplicate_closed_issues']}\n"
+        summary += f"  * {LOCALIZE[util.get_region()]['summary_fail_duplicate_closed_issues']}\n"
 
     # Total success
     if number_of_hits == 0:
         # TODO NEW Localization 'summary_success' | "SUCCESS: No new issues detected" #localization
-        summary += f"  * {LOCALIZE[get_region()]['summary_success']}\n"
+        summary += f"  * {LOCALIZE[util.get_region()]['summary_success']}\n"
 
     os.environ["TODOON_STATUS"] = "finished"  # todoon
     os.environ["TODOON_PROGRESS"] = "100.0"  # todoon
@@ -966,7 +660,7 @@ def todoon(  # todoon
         number_of_closed_issues
     )
 
-    print_wrap(log_level=log_level, msg_level=LOG_LEVEL_SUMMARY_ONLY,
+    util.print_wrap(log_level=log_level, msg_level=util.LOG_LEVEL_SUMMARY_ONLY,
                msg=summary, file=sys.stderr)
 
     # Fail if any hits were found and we are not in silent mode
@@ -1021,7 +715,7 @@ def todo_ignore_util(  # todoon
                             if len(line) > 0:
                                 output.append(line)
                 except FileNotFoundError:
-                    print(LOCALIZE[get_region()]["warning_file_does_not_exist"], _path, file=sys.stderr)
+                    print(LOCALIZE[util.get_region()]["warning_file_does_not_exist"], _path, file=sys.stderr)
 
     try:
         with open(todoignore_path, access_mode) as target:  # todoon
@@ -1034,11 +728,11 @@ def todo_ignore_util(  # todoon
                 target.write(f"{line}\n")
     except FileExistsError:
         print(
-            LOCALIZE[get_region()]["error_file_already_exists"], todoignore_path, file=sys.stderr  # todoon
+            LOCALIZE[util.get_region()]["error_file_already_exists"], todoignore_path, file=sys.stderr  # todoon
         )
         exit(1)
 
-    print(LOCALIZE[get_region()]["general_done"])
+    print(LOCALIZE[util.get_region()]["general_done"])
 
 
 def typer_todoon():  # todoon
